+++ src/compas_rhino/conversions/geometry.py
@@ -78,19 +78,7 @@ def frame_to_rhino_plane(frame):
     return RhinoPlane(point_to_rhino(frame.point), vector_to_rhino(frame.xaxis), vector_to_rhino(frame.yaxis))
 
 
-def frame_to_rhino(frame):
-    """Convert a COMPAS frame to a Rhino plane.
-
-    Parameters
-    ----------
-    frame : :class:`compas.geometry.Frame`
-
-    Returns
-    -------
-    :rhino:`Rhino.Geometry.Plane`
-
-    """
-    return RhinoPlane(point_to_rhino(frame[0]), vector_to_rhino(frame[1]), vector_to_rhino(frame[2]))
+frame_to_rhino = frame_to_rhino_plane
 
 
 def polygon_to_rhino(polygon):
@@ -175,6 +163,9 @@ def plane_to_compas_frame(plane):
         vector_to_compas(plane.XAxis),
         vector_to_compas(plane.YAxis),
     )
+
+
+frame_to_compas = plane_to_compas_frame
 
 
 def polygon_to_compas(polygon):