+++ tests/compas/scene/test_scene.py
@@ -1,6 +1,9 @@
 import pytest  # noqa: F401
 
 import compas
+from compas.scene import context
+from compas.scene import register
+from compas.scene import build_scene_object
 from compas.scene import SceneObject
 from compas.scene import NoSceneObjectContextError
 
@@ -8,16 +11,15 @@ from compas.scene import NoSceneObjectContextError
 if not compas.IPY:
 
     @pytest.fixture(autouse=True)
-    def reset_sceneobjects():
+    def reset_sceneobjects_registration():
         # before each test
         yield
         # after each test, reset scene objects
-        SceneObject.ITEM_SCENEOBJECT.clear()
-        SceneObject._SceneObject__SCENEOBJECTS_REGISTERED = False  # type: ignore
+        context.ITEM_SCENEOBJECT.clear()
 
 
 def register_fake_context():
-    SceneObject.register(FakeItem, FakeSceneObject, context="fake")
+    register(FakeItem, FakeSceneObject, context="fake")
 
 
 class FakeSceneObject(SceneObject):
@@ -39,52 +41,50 @@ class FakeSubItem(FakeItem):
 
 
 def test_get_sceneobject_cls_with_orderly_registration():
-    SceneObject.register(FakeItem, FakeSceneObject, context="fake")
-    SceneObject.register(FakeSubItem, FakeSubSceneObject, context="fake")
+    register(FakeItem, FakeSceneObject, context="fake")
+    register(FakeSubItem, FakeSubSceneObject, context="fake")
     item = FakeItem()
-    sceneobject = SceneObject(item, context="fake")
+    sceneobject = build_scene_object(item, context="fake")
     assert isinstance(sceneobject, FakeSceneObject)
 
     item = FakeSubItem()
-    sceneobject = SceneObject(item, context="fake")
+    sceneobject = build_scene_object(item, context="fake")
     assert isinstance(sceneobject, FakeSubSceneObject)
 
 
 def test_get_sceneobject_cls_with_out_of_order_registration():
-    SceneObject.register(FakeSubItem, FakeSubSceneObject, context="fake")
-    SceneObject.register(FakeItem, FakeSceneObject, context="fake")
+    register(FakeSubItem, FakeSubSceneObject, context="fake")
+    register(FakeItem, FakeSceneObject, context="fake")
     item = FakeItem()
-    sceneobject = SceneObject(item, context="fake")
+    sceneobject = build_scene_object(item, context="fake")
     assert isinstance(sceneobject, FakeSceneObject)
 
     item = FakeSubItem()
-    sceneobject = SceneObject(item, context="fake")
+    sceneobject = build_scene_object(item, context="fake")
     assert isinstance(sceneobject, FakeSubSceneObject)
 
 
 if not compas.IPY:
 
     def test_sceneobject_auto_context_discovery(mocker):
-        mocker.patch("compas.scene.SceneObject.register_scene_objects")
-        SceneObject.register_scene_objects.side_effect = register_fake_context
-        SceneObject._SceneObject__SCENEOBJECTS_REGISTERED = False  # type: ignore
+        register_fake_context()
 
         item = FakeItem()
-        sceneobject = SceneObject(item)
+        sceneobject = build_scene_object(item)
 
         assert isinstance(sceneobject, FakeSceneObject)
 
     def test_sceneobject_auto_context_discovery_viewer(mocker):
-        mocker.patch("compas.scene.sceneobject.is_viewer_open", return_value=True)
-        SceneObject.ITEM_SCENEOBJECT["Viewer"] = {FakeItem: FakeSceneObject}
+        mocker.patch("compas.scene.context.is_viewer_open", return_value=True)
+        context.ITEM_SCENEOBJECT["Viewer"] = {FakeItem: FakeSceneObject}
 
         item = FakeSubItem()
-        sceneobject = SceneObject(item)
+        sceneobject = build_scene_object(item)
 
         assert isinstance(sceneobject, FakeSceneObject)
 
     def test_sceneobject_auto_context_discovery_viewer_priority(mocker):
-        mocker.patch("compas.scene.sceneobject.is_viewer_open", return_value=True)
+        mocker.patch("compas.scene.context.is_viewer_open", return_value=True)
 
         class FakeViewerSceneObject(FakeSceneObject):
             pass
@@ -92,19 +92,19 @@ if not compas.IPY:
         class FakePlotterSceneObject(FakeSceneObject):
             pass
 
-        SceneObject.ITEM_SCENEOBJECT["Viewer"] = {FakeItem: FakeViewerSceneObject}
-        SceneObject.ITEM_SCENEOBJECT["Plotter"] = {FakeItem: FakePlotterSceneObject}
+        context.ITEM_SCENEOBJECT["Viewer"] = {FakeItem: FakeViewerSceneObject}
+        context.ITEM_SCENEOBJECT["Plotter"] = {FakeItem: FakePlotterSceneObject}
 
         item = FakeSubItem()
-        sceneobject = SceneObject(item)
+        sceneobject = build_scene_object(item)
 
         assert isinstance(sceneobject, FakeViewerSceneObject)
 
     def test_sceneobject_auto_context_discovery_no_context(mocker):
-        mocker.patch("compas.scene.sceneobject.is_viewer_open", return_value=False)
-        mocker.patch("compas.scene.sceneobject.compas.is_grasshopper", return_value=False)
-        mocker.patch("compas.scene.sceneobject.compas.is_rhino", return_value=False)
+        mocker.patch("compas.scene.context.is_viewer_open", return_value=False)
+        mocker.patch("compas.scene.context.compas.is_grasshopper", return_value=False)
+        mocker.patch("compas.scene.context.compas.is_rhino", return_value=False)
 
         with pytest.raises(NoSceneObjectContextError):
             item = FakeSubItem()
-            _ = SceneObject(item)
+            _ = build_scene_object(item)